custom_plugin_simple_example/requirements.txt
Pillow==7.2.0
pycuda
numpy
tensorflow-gpu
tensorrt
uf
graphsurgeon